minor simplify
