+++ samples/csharp/chatapp-net6/Program.cs
@@ -1,4 +1,5 @@
 using Microsoft.Azure.WebPubSub.AspNetCore;
+using Microsoft.Azure.WebPubSub.Common;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -13,7 +14,6 @@ if (app.Environment.IsDevelopment())
     app.UseDeveloperExceptionPage();
 }
 
-app.UseDefaultFiles();
 app.UseStaticFiles();
 app.UseRouting();
 
@@ -35,3 +35,25 @@ app.UseEndpoints(endpoints =>
 });
 
 app.Run();
+
+sealed class SampleChatHub : WebPubSubHub
+{
+    private readonly WebPubSubServiceClient<SampleChatHub> _serviceClient;
+
+    public SampleChatHub(WebPubSubServiceClient<SampleChatHub> serviceClient)
+    {
+        _serviceClient = serviceClient;
+    }
+
+    public override async Task OnConnectedAsync(ConnectedEventRequest request)
+    {
+        await _serviceClient.SendToAllAsync($"[SYSTEM] {request.ConnectionContext.UserId} joined.");
+    }
+
+    public override async ValueTask<UserEventResponse> OnMessageReceivedAsync(UserEventRequest request, CancellationToken cancellationToken)
+    {
+        await _serviceClient.SendToAllAsync($"[{request.ConnectionContext.UserId}] {request.Data}");
+
+        return request.CreateResponse($"[SYSTEM] ack.");
+    }
+}
